refactor(editor): remove URL sharing feature from editor template
Removed the section that displayed the shareable URL for the page being edited. This change simplifies the user interface by eliminating extraneous information that was not central to the editing experience, focusing instead on the task at hand. No functionality is affected aside from the removal of the URL display.